zip fix for windows build

+++ zipdist.js
@@ -19,7 +19,10 @@ class TestMe {
 
     // zip has circular problem so we do it in stages
     // delete old directorty
-    fs.rmdirSync(endDownLoadFolder, { recursive: true, force: true })
+    if (fs.existsSync(endDownLoadFolder)) {
+      fs.rmdirSync(endDownLoadFolder, { recursive: true, force: true })
+    }
+    console.log('removed')
     await zip.zip(zipFromFolder, initialZipFile)
     ensureDirSync(endDownLoadFolder)
     fs.renameSync(initialZipFile, endDownLoadFile)
